+++ rare_cases/case_try_except.py
@@ -1,6 +1,0 @@
-# This is not unused import, but it is unused import according to unimport.
-
-try:
-    import django
-except ImportError:
-    print("install django")